refactor: merge input and mode files
