+++ internal/tui/mode.go
@@ -1,6 +1,7 @@
 package tui
 
 import (
+	"github.com/Broderick-Westrope/tetrigo/internal/data"
 	tea "github.com/charmbracelet/bubbletea"
 )
 
@@ -42,4 +43,56 @@ var modeToStrMap = map[Mode]string{
 
 func (m Mode) String() string {
 	return modeToStrMap[m]
+}
+
+// SwitchModeInput values --------------------------------------------------
+
+type SingleInput struct {
+	Mode       Mode
+	Level      int
+	PlayerName string
+}
+
+func NewSingleInput(mode Mode, level int, playerName string) *SingleInput {
+	return &SingleInput{
+		Mode:       mode,
+		Level:      level,
+		PlayerName: playerName,
+	}
+}
+
+func (in *SingleInput) isSwitchModeInput() {}
+
+type MenuInput struct {
+}
+
+func NewMenuInput() *MenuInput {
+	return &MenuInput{}
+}
+
+func (in *MenuInput) isSwitchModeInput() {}
+
+type LeaderboardInput struct {
+	GameMode string
+	NewEntry *data.Score
+}
+
+func NewLeaderboardInput(gameMode string, opts ...func(input *LeaderboardInput)) *LeaderboardInput {
+	in := &LeaderboardInput{
+		GameMode: gameMode,
+	}
+
+	for _, opt := range opts {
+		opt(in)
+	}
+
+	return in
+}
+
+func (in *LeaderboardInput) isSwitchModeInput() {}
+
+func WithNewEntry(entry *data.Score) func(input *LeaderboardInput) {
+	return func(in *LeaderboardInput) {
+		in.NewEntry = entry
+	}
 }
